Refactor the database methods into a database manager
Move out the database logic into a dedicated db manager to separate concerns
